% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleList.R
\name{antDesignXBubbleListOutput}
\alias{antDesignXBubbleListOutput}
\title{Bubble List Widget — Read-only chat message list}
\usage{
antDesignXBubbleListOutput(outputId, width = "100\%", height = "500px", ...)
}
\arguments{
\item{outputId}{Output ID.}

\item{width, height}{CSS dimensions.}

\item{...}{Passed to [htmlwidgets::shinyWidgetOutput()].}
}
\description{
Bubble List Widget — Read-only chat message list
}
