% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antDesignX.R
\name{antDesignXOutput}
\alias{antDesignXOutput}
\title{Ant Design X Chat UI Output}
\usage{
antDesignXOutput(outputId, width = "100\%", height = "600px", ...)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Width and height of the widget (CSS values).}

\item{...}{Additional arguments passed to [htmlwidgets::shinyWidgetOutput()].}
}
\value{
An HTML output element.
}
\description{
Creates an htmlwidget output placeholder for the Ant Design X chat widget.
}
