% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welcome.R
\name{antDesignXWelcomeOutput}
\alias{antDesignXWelcomeOutput}
\title{Welcome Widget — Onboarding screen}
\usage{
antDesignXWelcomeOutput(outputId, width = "100\%", height = "auto", ...)
}
\arguments{
\item{outputId}{Output ID.}

\item{width, height}{CSS dimensions.}

\item{...}{Passed to [htmlwidgets::shinyWidgetOutput()].}
}
\description{
Welcome Widget — Onboarding screen
}
