% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antDesignX.R
\name{renderAntDesignX}
\alias{renderAntDesignX}
\title{Render an Ant Design X Widget}
\usage{
renderAntDesignX(config = list(), outputId = NULL)
}
\arguments{
\item{config}{Optional named list of configuration options.}

\item{outputId}{The output ID used in [antDesignXOutput()].}
}
\value{
A render function suitable for assigning to `output[[outputId]]`.
}
\description{
Server-side render function for [antDesignXOutput()].
}
