% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{renderAntDesignXAttachments}
\alias{renderAntDesignXAttachments}
\title{Render Attachments Widget}
\usage{
renderAntDesignXAttachments(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{A list with: `inputId`, `maxCount` (default 5), `multiple` (default TRUE),
`accept` (e.g. `".png,.jpg"`), `placeholderTitle`, `placeholderDescription`,
`overflow` (`"wrap"`, `"scrollX"`, `"scrollY"`).}

\item{env, quoted}{Passed to [shiny::exprToFunction()].}
}
\description{
Render Attachments Widget
}
