% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{renderAntDesignXPrompts}
\alias{renderAntDesignXPrompts}
\title{Render Prompts Widget}
\usage{
renderAntDesignXPrompts(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{A list with: `inputId` (string), `items` (list of `list(key, label, description)`),
`title` (string), `vertical` (logical), `wrap` (logical).}

\item{env, quoted}{Passed to [shiny::exprToFunction()].}
}
\description{
Render Prompts Widget
}
