% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sender.R
\name{renderAntDesignXSender}
\alias{renderAntDesignXSender}
\title{Render Sender Widget}
\usage{
renderAntDesignXSender(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{A list with: `inputId`, `placeholder` (string), `loading` (logical),
`allowSpeech` (logical), `submitType` (`"enter"` or `"shiftEnter"`).}

\item{env, quoted}{Passed to [shiny::exprToFunction()].}
}
\description{
Render Sender Widget
}
