% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestion.R
\name{renderAntDesignXSuggestion}
\alias{renderAntDesignXSuggestion}
\title{Render Suggestion Widget}
\usage{
renderAntDesignXSuggestion(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{A list with: `inputId`, `items` (list of `list(value, label, description)`),
`placeholder` (string), `block` (logical, default TRUE).}

\item{env, quoted}{Passed to [shiny::exprToFunction()].}
}
\description{
Render Suggestion Widget
}
