% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/think.R
\name{renderAntDesignXThink}
\alias{renderAntDesignXThink}
\title{Render Think Widget}
\usage{
renderAntDesignXThink(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{A list with: `content` (string), `title` (string),
`loading` (logical), `defaultExpanded` (logical).}

\item{env, quoted}{Passed to [shiny::exprToFunction()].}
}
\description{
Render Think Widget
}
