% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welcome.R
\name{renderAntDesignXWelcome}
\alias{renderAntDesignXWelcome}
\title{Render Welcome Widget}
\usage{
renderAntDesignXWelcome(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{A list with: `title` (string), `description` (string),
`variant` (`"filled"` or `"borderless"`).}

\item{env, quoted}{Passed to [shiny::exprToFunction()].}
}
\description{
Render Welcome Widget
}
