% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCard.R
\name{xcard_update_components}
\alias{xcard_update_components}
\title{Build A2UI updateComponents command}
\usage{
xcard_update_components(surface_id, components)
}
\arguments{
\item{surface_id}{Surface ID.}

\item{components}{List of component definitions. Each: list with `id`, `component`,
and any props.}
}
\description{
Build A2UI updateComponents command
}
