% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCard.R
\name{xcard_create_surface}
\alias{xcard_create_surface}
\title{Build A2UI createSurface command}
\usage{
xcard_create_surface(surface_id, catalog_id = "shiny-default")
}
\arguments{
\item{surface_id}{Surface ID string.}

\item{catalog_id}{Catalog ID (default `"shiny-default"`).}
}
\description{
Build A2UI createSurface command
}
